reworked cmake dependency handling

+++ qc-core/include/qc-core/record-allocator.hpp
@@ -106,6 +106,6 @@ namespace qc::memory
 
       private:
 
-        u64 _listI{};
+        u32 _listI{};
     };
 }
